+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppRootProps } from '@grafana/data';
-import { Routes } from 'Components/Routes';
+import { Routes } from 'Routes';
 
 export class App extends React.PureComponent<AppRootProps> {
   render() {
