+++ src/Components/MainScenes/ServiceSceneTabs/Patterns/FilterByPatternsButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { SceneObjectState, SceneObjectBase, SceneComponentProps, sceneGraph } from '@grafana/scenes';
 import { Button } from '@grafana/ui';
-import { IndexScene } from 'Components/Index/IndexScene';
+import { IndexScene } from 'Components/MainScenes/IndexScene';
 
 export interface FilterByPatternsButtonState extends SceneObjectState {
   pattern: string;
