+++ src/module.ts
@@ -1,4 +1,4 @@
 import { AppPlugin } from '@grafana/data';
-import { App } from 'Components/App';
+import { App } from 'App';
 
 export const plugin = new AppPlugin<{}>().setRootPage(App);
