+++ src/services/scenes.ts
@@ -8,7 +8,7 @@ import {
   SceneObjectUrlValues,
   SceneTimeRange,
 } from '@grafana/scenes';
-import { IndexScene } from 'Components/Index/IndexScene';
+import { IndexScene } from 'Components/MainScenes/IndexScene';
 import { VAR_DATASOURCE_EXPR, LOG_STREAM_SELECTOR_EXPR, VAR_FILTERS, ALL_VARIABLE_VALUE } from './variables';
 import { EXPLORATIONS_ROUTE } from './routing';
