--- src/services/types.ts
@@ -1,8 +1,0 @@
-export type DetectedLabel = {
-  label: string;
-  cardinality: number;
-};
-
-export type DetectedLabelsResponse = {
-  detectedLabels: DetectedLabel[];
-};
